+++ astor/code_gen.py
@@ -18,6 +18,7 @@ this code came from here (in 2012):
 """
 
 import ast
+import inspect
 import math
 import sys
 
@@ -53,10 +54,10 @@ def to_source(node, indent_with=' ' * 4, add_line_information=False,
     """
     if source_generator_class is None:
         source_generator_class = SourceGenerator
-    elif not isinstance(source_generator_class, SourceGenerator):
+    elif not inspect.isclass(source_generator_class):
+        raise TypeError('source_generator_class should be a class')
+    elif not issubclass(source_generator_class, SourceGenerator):
         raise TypeError('source_generator_class should be a subclass of SourceGenerator')
-    elif not callable(source_generator_class):
-        raise TypeError('source_generator_class should be a callable')
     generator = source_generator_class(
         indent_with, add_line_information, pretty_string)
     generator.visit(node)