+++ docs/changelog.rst
@@ -15,6 +15,14 @@ Bug fixes
 .. _`Issue 153`: https://github.com/berkerpeksag/astor/issues/153
 .. _`PR 155`: https://github.com/berkerpeksag/astor/pull/155
 
+* Fixed :func:`astor.to_source` incorrectly checking whether
+  *source_generator_class* is a subclass of :class:`astor.code_gen.SourceGenerator`.
+  (Reported by Yu-Chia "Hank" Liu in `Issue 158`_ and fixed by Will Crichton in `PR 164`_.)
+
+.. _`Issue 158`: https://github.com/berkerpeksag/astor/issues/158
+.. _`PR 164`: https://github.com/berkerpeksag/astor/pull/164
+
+
 0.8.0 - 2019-05-19
 ------------------